+++ previous-versions/assess/src/codatassess/models/operations/get_accounts_for_enhanced_profit_and_loss.py
@@ -29,7 +29,7 @@ class GetAccountsForEnhancedProfitAndLossResponse:
     enhanced_report: Optional[shared_enhancedreport.EnhancedReport] = dataclasses.field(default=None)
     r"""OK"""
     error_message: Optional[shared_errormessage.ErrorMessage] = dataclasses.field(default=None)
-    r"""Your API request was not properly authorized."""
+    r"""The request made is not valid."""
     raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
     r"""Raw HTTP response; suitable for custom response parsing"""