+++ previous-versions/assess/src/codatassess/models/operations/get_data_integrity_summaries.py
@@ -28,7 +28,7 @@ class GetDataIntegritySummariesResponse:
     status_code: int = dataclasses.field()
     r"""HTTP response status code for this operation"""
     error_message: Optional[shared_errormessage.ErrorMessage] = dataclasses.field(default=None)
-    r"""Your API request was not properly authorized."""
+    r"""Your `query` parameter was not correctly formed"""
     raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
     r"""Raw HTTP response; suitable for custom response parsing"""
     summaries: Optional[shared_summaries.Summaries] = dataclasses.field(default=None)