+++ previous-versions/assess/RELEASES.md
@@ -19,3 +19,13 @@ Based on:
 - [python v0.33.0] previous-versions/assess
 ### Releases
 - [PyPI v0.33.0] https://pypi.org/project/codat-assess/0.33.0 - previous-versions/assess
+
+## 2023-10-30 01:23:46
+### Changes
+Based on:
+- OpenAPI Doc 1.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Assess.yaml
+- Speakeasy CLI 1.109.0 (2.173.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v0.33.1] previous-versions/assess
+### Releases
+- [PyPI v0.33.1] https://pypi.org/project/codat-assess/0.33.1 - previous-versions/assess